+++ packages/bakit/src/lib/client/helpers/index.ts
@@ -1,0 +1,10 @@
+import type { Client } from "../index.js";
+import { createChannelHelpers } from "./channel.js";
+import { createUserHelpers } from "./user.js";
+
+export function createHelpers(client: Client) {
+	return {
+		user: createUserHelpers(client),
+		channel: createChannelHelpers(client),
+	};
+}